refactor(front): use the new centralized version
