+++ front/components/assistant/conversation/input_bar/editor/MentionComponent.tsx
@@ -1,21 +1,7 @@
-import {
-  ChatBubbleBottomCenterTextIcon,
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-  EyeIcon,
-  TooltipContent,
-  TooltipProvider,
-  TooltipRoot,
-  TooltipTrigger,
-} from "@dust-tt/sparkle";
 import { NodeViewWrapper } from "@tiptap/react";
-import { useRouter } from "next/router";
 import React from "react";
 
-import { useURLSheet } from "@app/hooks/useURLSheet";
-import { getConversationRoute, setQueryParam } from "@app/lib/utils/router";
+import { MentionDisplay } from "@app/lib/mentions";
 import type { WorkspaceType } from "@app/types";
 
 interface MentionComponentProps {
@@ -32,55 +18,22 @@ interface MentionComponentProps {
 }
 
 export const MentionComponent = ({ node, owner }: MentionComponentProps) => {
-  const router = useRouter();
-  const { onOpenChange: onOpenChangeAgentModal } = useURLSheet("agentDetails");
-
-  const { id: agentSId, label: agentName, description } = node.attrs;
-
-  const handleStartConversation = async () => {
-    if (!owner) {
-      return;
-    }
-    await router.push(
-      getConversationRoute(owner.sId, "new", `agent=${agentSId}`)
-    );
-  };
-
-  const handleSeeDetails = () => {
-    onOpenChangeAgentModal(true);
-    setQueryParam(router, "agentDetails", agentSId);
-  };
-
-  const trigger = (
-    <DropdownMenuTrigger asChild>
-      <span className="inline-block cursor-pointer font-medium text-highlight-500">
-        @{agentName}
-      </span>
-    </DropdownMenuTrigger>
-  );
+  const { id, label, description, pictureUrl, type } = node.attrs;
 
   return (
     <NodeViewWrapper className="inline-flex">
-      <TooltipProvider>
-        <DropdownMenu>
-          <TooltipRoot>
-            <TooltipTrigger asChild>{trigger}</TooltipTrigger>
-            {description && <TooltipContent>{description}</TooltipContent>}
-          </TooltipRoot>
-          <DropdownMenuContent side="top" align="start">
-            <DropdownMenuItem
-              onClick={handleStartConversation}
-              icon={ChatBubbleBottomCenterTextIcon}
-              label={`New conversation with @${agentName}`}
-            />
-            <DropdownMenuItem
-              onClick={handleSeeDetails}
-              icon={EyeIcon}
-              label={`About @${agentName}`}
-            />
-          </DropdownMenuContent>
-        </DropdownMenu>
-      </TooltipProvider>
+      <MentionDisplay
+        mention={{
+          id,
+          label,
+          description: description ?? "",
+          pictureUrl: pictureUrl ?? "",
+          type,
+        }}
+        interactive={!!owner}
+        owner={owner}
+        showTooltip={true}
+      />
     </NodeViewWrapper>
   );
 };
