+++ front/lib/api/assistant/conversation/mentions.ts
@@ -6,11 +6,11 @@ import {
   Message,
 } from "@app/lib/models/assistant/conversation";
 import { generateRandomModelSId } from "@app/lib/resources/string_ids";
+import type { MentionType } from "@app/types";
 import type {
   AgentMessageType,
   ConversationType,
   LightAgentConfigurationType,
-  MentionType,
   UserMessageType,
   WorkspaceType,
 } from "@app/types";
