+++ front/types/assistant/conversation.ts
@@ -3,6 +3,7 @@ import type { ActionGeneratedFileType } from "@app/lib/actions/types";
 import type { AllSupportedWithDustSpecificFileContentType } from "@app/types";
 import type { ContentFragmentType } from "@app/types";
 import type { ModelId } from "@app/types";
+import type { MentionType } from "@app/types";
 import type { AgentMCPActionWithOutputType } from "@app/types/actions";
 
 import type { UserType, WorkspaceType } from "../user";
@@ -13,26 +14,7 @@ import type {
 } from "./agent";
 import type { AgentContentItemType } from "./agent_message_content";
 
-/**
- * Mentions
- */
-
-export type AgentMention = {
-  configurationId: string;
-};
-
-export type UserMention = {
-  type: "user";
-  userId: string;
-};
-
-export type MentionType = AgentMention | UserMention;
-
 export type MessageVisibility = "visible" | "deleted";
-
-export function isAgentMention(arg: MentionType): arg is AgentMention {
-  return (arg as AgentMention).configurationId !== undefined;
-}
 
 export type ConversationMessageReactions = {
   messageId: string;
