+++ front/types/assistant/mentions.ts
